Move global options to external parameter. Dual visualization.
The options of the visualization are not stored as constants in the
class but are dynamically injected when the class is created.
There are now two visualizations shown side-by-side: the Marey
diagram and the interactive map.
The vertical line that separates the two visualizations can be
moved left or right modifying the percentage parameter.
The Marey diagram is not yet implemented.