Initial Marey domain based on time units.
Previously, the initially selected domain in the marey diagram was
1/4 of the total domain. Since the total domain is defined by the
start and end times of the selected journey pattern(s), this meant
that if a journey pattern had trips all day long (24hrs) the initial
domain was the first 6hrs while if a journey pattern had trips only for
two hours the first 30min were selected.
Now, it's always the first two hours of the total domain that are
selected. If total domain < 2hrs, the entire domain is selected.