(Marey): make maximum zoom level based on total domain.
Previously, the maximum zoom factor in the Marey diagram was set
absolutely. This meant that for some datasets the maximum zoom
was at the second level while for other ones it was at the minutes
level.
Now, the zoom factor is based on the total time in the dataset. This
means that at the maximum zoom level the granularity is always seconds.